+++ internal/mode/usage/markdown/relationship.go
@@ -1,6 +1,8 @@
 package markdown
 
-import "github.com/dogmatiq/ferrite/variable"
+import (
+	"github.com/dogmatiq/ferrite/variable"
+)
 
 func (r *specRenderer) renderSeeAlso() {
 	relationships := variable.FilterRelationships[variable.RefersTo](r.spec)