+++ internal/mode/usage/markdown/spec.go
@@ -124,6 +124,21 @@ func (r *specRenderer) renderPrimaryRequirement(f string, v ...any) {
 		text = "⚠️ The `%s` variable is **deprecated**; its use is **NOT RECOMMENDED** as it may be removed in a future version."
 		args = []any{r.spec.Name()}
 
+		relationships := variable.FilterRelationships[variable.IsSupersededBy](r.spec)
+		if len(relationships) != 0 {
+			for i, rel := range relationships {
+				if i == len(relationships)-1 {
+					text += " and"
+				} else if i > 0 {
+					text += ","
+				}
+
+				text += " " + r.ren.linkToSpec(rel.SupersededBy)
+			}
+
+			text += " **SHOULD** be used instead."
+		}
+
 		if req != "" {
 			text += " If defined, the value %s."
 			args = append(args, req)